# build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "patlytics_infringescan_backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "Error"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"data": {
										"properties": {
											"password": {
												"type": "string"
											},
											"email": {
												"type": "string"
											},
											"last_name": {
												"type": "string"
											},
											"first_name": {
												"type": "string"
											}
										},
										"required": [
											"password",
											"email",
											"last_name",
											"first_name"
										],
										"type": "object"
									}
								},
								"required": [
									"data"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "Error"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"data": {
										"properties": {
											"password": {
												"type": "string"
											},
											"email": {
												"type": "string"
											}
										},
										"required": [
											"password",
											"email"
										],
										"type": "object"
									}
								},
								"required": [
									"data"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/refreshToken": {
			"post": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "Error"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"refreshToken": {
										"type": "string"
									}
								},
								"required": [
									"refreshToken"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/company/list": {
			"get": {
				"operationId": "GetCompanies",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "Error"
					}
				},
				"tags": [
					"Company"
				],
				"security": [],
				"parameters": []
			}
		},
		"/company/{company_id}/products": {
			"get": {
				"operationId": "GetCompanyProducts",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "Error"
					}
				},
				"tags": [
					"Company"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "company_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/comparison/getComparisonByUuid/{uuid}": {
			"get": {
				"operationId": "GetComparisonByUuid",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "Error"
					}
				},
				"tags": [
					"Comparison"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/comparison/getOrCreateComparison": {
			"post": {
				"operationId": "GetOrCreateComparison",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "Error"
					}
				},
				"tags": [
					"Comparison"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"data": {
										"properties": {
											"publication_number": {
												"type": "string"
											},
											"company_uuid": {
												"type": "string"
											}
										},
										"required": [
											"publication_number",
											"company_uuid"
										],
										"type": "object"
									}
								},
								"required": [
									"data"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/comparison/getOrCreateComparisonByUser": {
			"post": {
				"operationId": "GetOrCreateComparisonByUser",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "Error"
					}
				},
				"tags": [
					"Comparison"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"data": {
										"properties": {
											"publication_number": {
												"type": "string"
											},
											"company_uuid": {
												"type": "string"
											}
										},
										"required": [
											"publication_number",
											"company_uuid"
										],
										"type": "object"
									}
								},
								"required": [
									"data"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/patent/list": {
			"get": {
				"operationId": "GetPatents",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "Error"
					}
				},
				"tags": [
					"Patent"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "keyword",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/patent/{publication_number}/detail": {
			"get": {
				"operationId": "GetPatentByPublicationNumber",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "Error"
					}
				},
				"tags": [
					"Patent"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "publication_number",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/userComparison/user/list": {
			"get": {
				"operationId": "GetListByUserUuid",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "Error"
					}
				},
				"tags": [
					"UserComparison"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/user": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "Error"
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}